fix(frontend): resolve TypeScript compilation errors
- Fixed performanceMonitor.ts: Replace navigationStart with startTime for modern Performance API
- Fixed LazyComponentLoader.tsx: Resolved generic type constraints and added proper type annotations
- Fixed LazyComponents/index.ts: Removed JSX syntax from .ts file and updated component imports
- Fixed PerformanceProvider.tsx: Corrected severity type and added missing type annotations
- Fixed advancedCache.ts: Resolved type mismatches for compression/decompression APIs
- Fixed performanceRegression.ts: Resolved iterator compatibility issues
- Fixed userExperienceTracker.ts: Resolved NodeList iteration issues

All TypeScript errors resolved, clean production build achieved.

+++ database_models.py
@@ -30,7 +30,7 @@ class Email(Base):
     __tablename__ = 'emails'
     
     # Primary identifiers
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)  # Use Integer for SQLite compatibility
     message_id = Column(BigInteger, unique=True, nullable=False, index=True)
     apple_document_id = Column(String(255), index=True)
